+++ src/server/base.js
@@ -65,6 +65,7 @@ const postgraphileOptions = {
   // ignoreIndexes: false,
   // showErrorStack: "json",
   // extendedErrors: ["hint", "detail", "errcode"],
+  extendedErrors: ['severity', 'code', 'detail', 'hint', 'position', 'internalPosition', 'internalQuery', 'where', 'schema', 'table', 'column', 'dataType', 'constraint', 'file', 'line', 'routine'],
   // // appendPlugins: [require("@graphile-contrib/pg-simplify-inflector")],
   // allowExplain(req) {
   //   // TODO: customise condition!